Add external API types to be published
