+++ api/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "@clangd/vscode-clangd",
+    "version": "0.0.0",
+    "description": "API for the llvm-vs-code-extensions.vscode-clangd VS Code extension",
+    "types": "vscode-clangd.d.ts",
+    "publishConfig": {
+        "registry": "https://npm.pkg.github.com"
+    },
+    "repository": "https://github.com/clangd/vscode-clangd",
+    "dependencies": {
+        "vscode-languageclient": "8.0.2"
+    }
+}
